python3_requirements.txt
jupyter==1.0.0
matplotlib==3.7.1
numpy==1.24.3
python-dotenv==1.0.0
requests==2.29.0